+++ examples/api-auth-jwt-cognito-user-pool/sst.config.ts
@@ -1,11 +1,14 @@
 import { SSTConfig } from "sst";
+import { ExampleStack } from "./stacks/ExampleStack";
 
 export default {
   config(_input) {
     return {
-      name: "@@app",
+      name: "api-auth-jwt-cognito-user-pool",
       region: "us-east-1",
     };
   },
-  stacks(app) {},
+  stacks(app) {
+    app.stack(ExampleStack);
+  }
 } satisfies SSTConfig;