+++ examples/create-sst-dynamo/packages/web/src/components/Navbar.tsx
@@ -1,6 +1,6 @@
 import {CgLink} from "react-icons/cg";
 import {Link, useNavigate} from "react-router-dom";
-import {useTypedMutation} from "@@@app/graphql/urql";
+import {useTypedMutation} from "@create-sst-dynamo/graphql/urql";
 import Button from "./Button";
 import styles from "./Navbar.module.css";