+++ packages/sst/src/constructs/Script.ts
@@ -11,12 +11,9 @@ import {
   FunctionProps,
   FunctionDefinition,
 } from "./Function.js";
-import {
-  SSTConstruct,
-  SSTConstructMetadata,
-  getFunctionRef,
-} from "./Construct.js";
+import { SSTConstruct, getFunctionRef } from "./Construct.js";
 import { Permissions } from "./util/permission.js";
+import { BindingResource } from "./util/binding.js";
 const __dirname = path.dirname(url.fileURLToPath(import.meta.url));
 
 export interface ScriptProps {
@@ -172,7 +169,7 @@ export class Script extends Construct implements SSTConstruct {
    * script.bind([STRIPE_KEY, bucket]);
    * ```
    */
-  public bind(constructs: SSTConstruct[]): void {
+  public bind(constructs: BindingResource[]): void {
     this.createFunction?.bind(constructs);
     this.updateFunction?.bind(constructs);
     this.deleteFunction?.bind(constructs);
@@ -314,7 +311,7 @@ export class Script extends Construct implements SSTConstruct {
   }
 
   /** @internal */
-  public getFunctionBinding() {
+  public getBindings() {
     return undefined;
   }
 }