+++ packages/sst/src/constructs/Service.ts
@@ -7,7 +7,11 @@ import { existsAsync } from "../util/fs.js";
 import { Colors } from "../cli/colors.js";
 
 import { Construct } from "constructs";
-import { Duration as CdkDuration, IgnoreMode } from "aws-cdk-lib/core";
+import {
+  Duration as CdkDuration,
+  DockerCacheOption,
+  IgnoreMode,
+} from "aws-cdk-lib/core";
 import {
   Role,
   Effect,
@@ -39,12 +43,13 @@ import { Secret } from "./Secret.js";
 import { useDeferredTasks } from "./deferred_task.js";
 import { Permissions, attachPermissionsToRole } from "./util/permission.js";
 import {
-  FunctionBindingProps,
-  bindEnvironment,
-  bindPermissions,
+  BindingProps,
   getParameterPath,
-  getReferencedSecrets,
-} from "./util/functionBinding.js";
+  getBindingEnvironments,
+  getBindingPermissions,
+  getBindingReferencedSecrets,
+  BindingResource,
+} from "./util/binding.js";
 import { useProject } from "../project.js";
 import { IVpc, Vpc } from "aws-cdk-lib/aws-ec2";
 import {
@@ -177,6 +182,7 @@ const supportedMemories = {
 export interface ServiceDomainProps extends DistributionDomainProps {}
 export interface ServiceCdkDistributionProps
   extends Omit<DistributionProps, "defaultBehavior"> {}
+export interface ServiceContainerCacheProps extends DockerCacheOption {}
 
 export interface ServiceProps {
   /**
@@ -326,7 +332,7 @@ export interface ServiceProps {
    * }
    * ```
    */
-  bind?: SSTConstruct[];
+  bind?: BindingResource[];
   /**
    * The customDomain for this service. SST supports domains that are hosted
    * either on [Route 53](https://aws.amazon.com/route53/) or externally.
@@ -414,6 +420,42 @@ export interface ServiceProps {
      * ```
      */
     buildArgs?: Record<string, string>;
+    /**
+     * SSH agent socket or keys to pass to the docker build command.
+     * Docker BuildKit must be enabled to use the ssh flag
+     * @default No --ssh flag is passed to the build command
+     * @example
+     * ```js
+     * container: {
+     *   buildSsh: "default"
+     * }
+     * ```
+     */
+    buildSsh?: string;
+    /**
+     * Cache from options to pass to the docker build command.
+     * [DockerCacheOption](https://docs.aws.amazon.com/cdk/api/v2/docs/aws-cdk-lib.aws_ecr_assets.DockerCacheOption.html)[].
+     * @default No cache from options are passed to the build command
+     * @example
+     * ```js
+     * container: {
+     *   cacheFrom: [{ type: 'registry', params: { ref: 'ghcr.io/myorg/myimage:cache' }}],
+     * }
+     * ```
+     */
+    cacheFrom?: ServiceContainerCacheProps[];
+    /**
+     * Cache to options to pass to the docker build command.
+     * [DockerCacheOption](https://docs.aws.amazon.com/cdk/api/v2/docs/aws-cdk-lib.aws_ecr_assets.DockerCacheOption.html)[].
+     * @default No cache to options are passed to the build command
+     * @example
+     * ```js
+     * container: {
+     *   cacheTo: { type: 'registry', params: { ref: 'ghcr.io/myorg/myimage:cache', mode: 'max', compression: 'zstd' }},
+     * }
+     * ```
+     */
+    cacheTo?: ServiceContainerCacheProps;
   };
   dev?: {
     /**
@@ -724,7 +766,7 @@ export class Service extends Construct implements SSTConstruct {
   }
 
   /** @internal */
-  public getFunctionBinding(): FunctionBindingProps {
+  public getBindings(): BindingProps {
     const app = this.node.root as App;
     return this.distribution
       ? {
@@ -767,7 +809,7 @@ export class Service extends Construct implements SSTConstruct {
    * service.bind([STRIPE_KEY, bucket]);
    * ```
    */
-  public bind(constructs: SSTConstruct[]): void {
+  public bind(constructs: BindingResource[]): void {
     this.devFunction?.bind(constructs);
     this.bindForService(constructs);
   }
@@ -1087,31 +1129,23 @@ export class Service extends Construct implements SSTConstruct {
     });
   }
 
-  private bindForService(constructs: SSTConstruct[]): void {
+  private bindForService(constructs: BindingResource[]): void {
     // Get referenced secrets
     const referencedSecrets: Secret[] = [];
-    constructs.forEach((c) =>
-      referencedSecrets.push(...getReferencedSecrets(c))
+    constructs.forEach((r) =>
+      referencedSecrets.push(...getBindingReferencedSecrets(r))
     );
 
-    [...constructs, ...referencedSecrets].forEach((c) => {
+    [...constructs, ...referencedSecrets].forEach((r) => {
       // Bind environment
-      const env = bindEnvironment(c);
+      const env = getBindingEnvironments(r);
       Object.entries(env).forEach(([key, value]) =>
         this.addEnvironmentForService(key, value)
       );
 
       // Bind permissions
-      const permissions = bindPermissions(c);
-      Object.entries(permissions).forEach(([action, resources]) =>
-        this.attachPermissionsForService([
-          new PolicyStatement({
-            actions: [action],
-            effect: Effect.ALLOW,
-            resources,
-          }),
-        ])
-      );
+      const policyStatements = getBindingPermissions(r);
+      this.attachPermissionsForService(policyStatements);
     });
   }
 
@@ -1196,6 +1230,31 @@ export class Service extends Construct implements SSTConstruct {
           ...Object.entries(build?.buildArgs || {}).map(
             ([k, v]) => `--build-arg ${k}=${v}`
           ),
+          ...(build?.buildSsh ? [`--ssh ${build.buildSsh}`] : []),
+          ...(build?.cacheFrom || []).map(
+            (v) =>
+              "--cache-from=" +
+              [
+                `type=${v.type}`,
+                ...(v.params
+                  ? Object.entries(v.params).map(([pk, pv]) => `${pk}=${pv}`)
+                  : []),
+              ].join(",")
+          ),
+          ...(build?.cacheTo
+            ? [
+                "--cache-to " +
+                  [
+                    `type=${build?.cacheTo.type}`,
+                    ...(build?.cacheTo?.params
+                      ? Object.entries(build?.cacheTo?.params).map(
+                          ([pk, pv]) => `${pk}=${pv}`
+                        )
+                      : []
+                    ).join(","),
+                  ],
+              ]
+            : []),
           this.props.path,
         ].join(" "),
         {
@@ -1221,6 +1280,9 @@ export class Service extends Construct implements SSTConstruct {
         architecture === "arm64" ? Platform.LINUX_ARM64 : Platform.LINUX_AMD64,
       file: dockerfile,
       buildArgs: build?.buildArgs,
+      buildSsh: build?.buildSsh,
+      cacheFrom: build?.cacheFrom,
+      cacheTo: build?.cacheTo,
       exclude: [".sst/dist", ".sst/artifacts"],
       ignoreMode: IgnoreMode.GLOB,
     });