+++ packages/sst/support/remix-site-function/edge-server.js
@@ -1,19 +1,11 @@
 // This is a custom Lambda@Edge handler which imports the Remix server
 // build and performs the Remix server rendering.
 
-// We have to ensure that our polyfills are imported prior to any other modules
-// which may depend on them;
-import { installGlobals } from "@remix-run/node";
-installGlobals();
-
 import {
   createRequestHandler as createNodeRequestHandler,
   readableStreamToString,
 } from "@remix-run/node";
 import { URL } from "url";
-
-// Import the server build that was produced by `remix build`;
-import * as remixServerBuild from "./index.js";
 
 /**
  * Common binary MIME types